Android: remove the 'empty static lib include trick' hack
The code was using a hack to import headers only through a empty lib. This
trick does not work on darwin hosts.

This patch remove this hack and links against the real libraries. However,
since header directories are not recursively detected on Android, we need to
manually add all necessary libraries (here: libicuuc needed by libxml2).

Change-Id: Ibce07b5a41d2b5cbcb3b5757724740756f0ef904
Signed-off-by: David Wagner <david.wagner@intel.com>